requirements.txt
Django
python-dotenv
python-decouple
psycopg2-binary
Pillow
django-crispy-forms
crispy-bootstrap5
django-debug-toolbar
django-widget-tweaks
